4 tasks done for lesson_01 upd

+++ lesson_01/task_03_(exercise_09).py
@@ -1,9 +1,9 @@
 # Упражнение из списка заданий под номером 9
 # Вводятся три разных числа. Найти, какое из них является средним (больше одного, но меньше другого).
 
-a = int(input('Type in first number'))
-b = int(input('Type in second number'))
-c = int(input('Type in third number'))
+a = float(input('Type in first number'))
+b = float(input('Type in second number'))
+c = float(input('Type in third number'))
 
 if a > b:
     if a > c:
